MissionResultScreen: -started refractoring
MenuScren:
-moved textButton style code initialization and declaration to GameStyles class

General:
-moved GameImages and GameSounds classes into their own directory - game_resources/
-added code for java target versions in build.gradle
-updated libgdx dependencies to 1.9.11

build.gradle(:core):
-updated box2DLightsVersion to 1.5
-updated json to 20200518

XmlMapReader - removed catch blocks for now

+++ core/src/com/tankzor/game/common_value/Dimension.java
@@ -12,7 +12,7 @@ public class Dimension {
     public static final float BASE_SCREEN_HEIGHT = 720.0f;
     public static float gameObjectScaleFactor = 2.0f;
     public static float buttonWidth = 500.0f;
-    public static Integer buttonHeight = 80.0f;
+    public static float buttonHeight = 80.0f;
     public static float buttonSpace = 10.0f;
     public static float smallFontScale = 1.0f;
     public static float normalFontScale = 1.1f;
